affichage list sur main
Affichage des listes d'ingrédients sur la page du main avec récupération sur firebase
