+++ src/app/components/main/main.component.html
@@ -7,13 +7,19 @@
       <div class="sidebar-background"></div>
       <div class="sidebar-list">
         <div class="MyList-header">
-          <h2 class="sidebar-title"><mat-icon>list</mat-icon> My List</h2>
-          <button class="MyListButton" mat-icon-button color="primary" (click)="addRecipe()">
+          <h2 class="sidebar-title"><mat-icon>list</mat-icon> My Lists</h2>
+          <button class="MyListButton" mat-icon-button color="primary" (click)="addList()">
             <mat-icon id="MyList-AddButton">add</mat-icon>
           </button>
         </div>
         <div class="sidebar-list-container">
           <!-- Content for My List -->
+          <div class="sidebar-card" *ngFor="let recipe of ingredientsList">
+            <div class="sidebar-recipe-name">{{ recipe.name }}</div>
+            <div class="sidebar-recipe-attributs">
+              <div>{{ recipe.ingredients.length }} ingredients</div>
+            </div>
+          </div>
         </div>
       </div>
       <div class="sidebar-list">
@@ -92,11 +98,12 @@
         <mat-form-field appearance="outline">
           <mat-label>Ingredients</mat-label>
           <mat-select [formControl]="selectedIngredientsList">
+            <mat-option [value]="">No List</mat-option>
             <mat-option
               *ngFor="let ingredient of ingredientsList"
-              [value]="ingredient.listName"
+              [value]="ingredient.name"
             >
-              {{ ingredient.listName }}
+              {{ ingredient.name }}
             </mat-option>
           </mat-select>
         </mat-form-field>
