+++ src/app/components/page-list-ingredient/page-list-ingredient.component.html
@@ -2,7 +2,7 @@
     <!-- Main Content -->
     <div class="main-content">
         <div class="header">
-            <h1 class="page-title">Création d'une liste</h1>
+          <h1 class="page-title">Creating a list</h1>
         </div>
        
         <!-- Wrapper pour les colonnes -->
@@ -83,8 +83,6 @@
                   </div>
                 }
               </div>
-              
-             
                 <div class="button">
                     <button class="cancel-button" [routerLink]="['/main']">Annuler</button>
                     <button class="save-button" (click)="saveList()">Enregistrer</button>
