+++ src/app/components/page-list-ingredient/page-list-ingredient.component.ts
@@ -102,5 +102,4 @@ export class PageListIngredientComponent implements OnInit {
       this.filteredIngredients.unshift(removedIngredient);
     }
   }
-  
 }
