+++ src/app/models/ingredient.model.ts
@@ -1,6 +1,14 @@
 export interface Ingredient {
+
   idIngredient: string;
   strIngredient: string;
   strDescription?: string;
   strType?: string;
+}
+
+export interface IngredientList {
+  id?: string;
+  name: string;
+  ingredients: Ingredient[];
+  createdAt: Date;
 }
